Overloaded BookRoomNew so that the user can book a series of weekly meetings, by specifying the number of weeks:
Booking[] BookRoomNew(User userObj, Room roomObj, String bDate, String bStart, String bEnd, int noOfWeeks)

This method now also checks if the room is already booked. if it is BookRoomNew returns Null. This check is implemented in the isRoomAvailable method
addWeek is a new helper method that takes a given date and adds one week this method is used by the new BookRoomNew method